fix(ui): explain the GitHub App install step on connection rows
Replace the bare "Install on GitHub" link with an always-visible hint
explaining that installing the app on an organization is what grants
access to its private repos. Shown on every surface a connection row
renders (creation wizard, sandbox configure, settings), for GitHub and
GitHub Enterprise, when the connection carries a GitHub App slug.
