+++ packages/ui/src/modules/connections/components/catalog-connection-row.tsx
@@ -10,7 +10,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 
-import { GithubAppInstallLink } from "./github-app-install-link.js";
+import { GithubAppInstallHint } from "./github-app-install-hint.js";
 
 export interface RowGrantControls {
   granted: boolean;
@@ -42,71 +42,75 @@ export function CatalogConnectionRow({
   return (
     <div
       data-testid={`catalog-connection-${connection.id}`}
-      className="flex items-center gap-2 px-4 py-3"
+      className="px-4 py-3"
     >
-      <div className="min-w-0 flex-1">
-        <div className="flex items-center gap-2">
-          <p className="truncate text-[15px] text-foreground">
-            {connection.name}
+      <div className="flex items-center gap-2">
+        <div className="min-w-0 flex-1">
+          <div className="flex items-center gap-2">
+            <p className="truncate text-[15px] text-foreground">
+              {connection.name}
+            </p>
+            {connection.status !== "active" && (
+              <StatusBadge status={connection.status} />
+            )}
+          </div>
+          <p className="truncate text-[14px] text-muted-foreground">
+            {subtitle}
           </p>
-          {connection.status !== "active" && (
-            <StatusBadge status={connection.status} />
-          )}
         </div>
-        <p className="truncate text-[14px] text-muted-foreground">{subtitle}</p>
-      </div>
-      <GithubAppInstallLink connection={connection} />
-      {grant &&
-        !grant.actionHidden &&
-        (grant.granted ? (
-          <span className="inline-flex h-[32px] shrink-0 items-center gap-1.5 rounded-full bg-muted px-3 text-[14px] text-foreground">
-            <Checkmark size={16} className="text-success" />
-            In this sandbox
-          </span>
-        ) : (
-          <Button
-            variant="outline"
-            className="h-[32px] shrink-0 px-3 text-[14px] font-normal"
-            onClick={() => grant.onToggle(true)}
-            data-testid={`catalog-add-${connection.id}`}
-          >
-            <Add size={16} />
-            Add to sandbox
-          </Button>
-        ))}
-      <DropdownMenu>
-        <DropdownMenuTrigger asChild>
-          <Button
-            variant="ghost"
-            size="icon-sm"
-            aria-label={`Actions for ${connection.name}`}
-            data-testid={`catalog-menu-${connection.id}`}
-          >
-            <OverflowMenuHorizontal size={16} />
-          </Button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent>
-          {grant?.granted && (
-            <DropdownMenuItem onSelect={() => grant.onToggle(false)}>
-              Remove from this sandbox
-            </DropdownMenuItem>
-          )}
-          {onManage && (
-            <DropdownMenuItem onSelect={onManage}>
-              Manage connections
-            </DropdownMenuItem>
-          )}
-          {onDelete && (
-            <DropdownMenuItem
-              tone="danger"
-              disabled={deleting}
-              onSelect={onDelete}
+        {grant &&
+          !grant.actionHidden &&
+          (grant.granted ? (
+            <span className="inline-flex h-[32px] shrink-0 items-center gap-1.5 rounded-full bg-muted px-3 text-[14px] text-foreground">
+              <Checkmark size={16} className="text-success" />
+              In this sandbox
+            </span>
+          ) : (
+            <Button
+              variant="outline"
+              className="h-[32px] shrink-0 px-3 text-[14px] font-normal"
+              onClick={() => grant.onToggle(true)}
+              data-testid={`catalog-add-${connection.id}`}
             >
-              Delete this connection
-            </DropdownMenuItem>
-          )}
-        </DropdownMenuContent>
-      </DropdownMenu>
+              <Add size={16} />
+              Add to sandbox
+            </Button>
+          ))}
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild>
+            <Button
+              variant="ghost"
+              size="icon-sm"
+              aria-label={`Actions for ${connection.name}`}
+              data-testid={`catalog-menu-${connection.id}`}
+            >
+              <OverflowMenuHorizontal size={16} />
+            </Button>
+          </DropdownMenuTrigger>
+          <DropdownMenuContent>
+            {grant?.granted && (
+              <DropdownMenuItem onSelect={() => grant.onToggle(false)}>
+                Remove from this sandbox
+              </DropdownMenuItem>
+            )}
+            {onManage && (
+              <DropdownMenuItem onSelect={onManage}>
+                Manage connections
+              </DropdownMenuItem>
+            )}
+            {onDelete && (
+              <DropdownMenuItem
+                tone="danger"
+                disabled={deleting}
+                onSelect={onDelete}
+              >
+                Delete this connection
+              </DropdownMenuItem>
+            )}
+          </DropdownMenuContent>
+        </DropdownMenu>
+      </div>
+      <GithubAppInstallHint connection={connection} />
     </div>
   );
 }
